% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeDmW}
\alias{ComputeDmW}
\title{compute D-alpha*W for CAR prior}
\usage{
ComputeDmW(A, sigma, alpha, thr_car)
}
\arguments{
\item{A}{symmetric matrix}

\item{sigma}{Kernel parameter}

\item{alpha}{Tunning parameter}

\item{thr_car}{Threashold to cut}
}
\description{
compute D-alpha*W for CAR prior
}
