% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeKernelParamLessMem}
\alias{ComputeKernelParamLessMem}
\title{Compute Gaussian kernel parameters}
\usage{
ComputeKernelParamLessMem(xin)
}
\arguments{
\item{xin}{spatial location}
}
\value{
A list
}
\description{
Compute Gaussian kernel parameters
}
