% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeTestQuantRcpp_RLSKAT}
\alias{ComputeTestQuantRcpp_RLSKAT}
\title{Compute the testing quantities for linear mixed model, float format}
\usage{
ComputeTestQuantRcpp_RLSKAT(cov_matin, Xdaggerin, covariatesin, norm_countsin)
}
\arguments{
\item{cov_matin}{Kernel matrix to be testedt}

\item{Xdaggerin}{The inverse matrix of covariates}

\item{covariatesin}{Covariates matrix}

\item{norm_countsin}{normalized count matrix}
}
\value{
A list
}
\description{
Compute the testing quantities for linear mixed model, float format
}
