% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialx.R
\name{DefaultAssay}
\alias{DefaultAssay}
\alias{DefaultAssay.SpatialX}
\alias{DefaultAssay<-}
\alias{DefaultAssay<-.SpatialX}
\title{Get and set default assay}
\usage{
DefaultAssay(object, ...)

\method{DefaultAssay}{SpatialX}(object, ...)

DefaultAssay(object, ...) <- value

\method{DefaultAssay}{SpatialX}(object, ...)

\method{DefaultAssay}{SpatialX}(object, ...) <- value
}
\arguments{
\item{object}{A SpatialX object}

\item{...}{Additional arguments passed to methods}

\item{value}{Name of the assay to set as default}
}
\value{
\code{DefaultAssay}: Returns the name of the current default assay

\code{DefaultAssay<-}: Returns the SpatialX object with updated default assay
}
\description{
These functions get and set the default (active) assay for SpatialX objects.
The default assay determines which assay is used by default in analysis and
visualization functions.
}
\details{
The default assay is used by many functions in the package when no specific
assay is provided. This allows users to seamlessly work with multiple assays
while focusing analysis on a particular data modality.
}
\examples{
\dontrun{
# Get the current default assay
current_assay <- DefaultAssay(spatialx_obj)
print(current_assay)

# Set a new default assay
DefaultAssay(spatialx_obj) <- "ATAC"

# Verify the change
DefaultAssay(spatialx_obj)  # Returns "ATAC"


# Switch between assays for different analyses
DefaultAssay(spatialx_obj) <- "RNA"
variable_genes <- FindVariableFeatures(spatialx_obj)

DefaultAssay(spatialx_obj) <- "Spatial"
spatial_genes <- FindSpatialGenes(spatialx_obj)
}


}
