% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialx.R
\name{dim.Assay}
\alias{dim.Assay}
\title{object-dimensions Number of cells and features for an Assay object}
\usage{
\method{dim}{Assay}(x, ...)
}
\arguments{
\item{x}{A SpatialX or Assay object}

\item{...}{Additional arguments passed to methods}
}
\value{
A numeric vector with the dimensions
}
\description{
object-dimensions Number of cells and features for an Assay object
}
