% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialx.R
\name{dim.SpatialX}
\alias{dim.SpatialX}
\title{object-dimensions Number of cells and features for the active assay}
\usage{
\method{dim}{SpatialX}(x, ...)
}
\arguments{
\item{x}{A SpatialX or Assay object}

\item{...}{Additional arguments passed to methods}
}
\value{
A numeric vector with the dimensions
}
\description{
object-dimensions Number of cells and features for the active assay
}
