% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialx.R
\name{dimnames.Assay}
\alias{dimnames.Assay}
\title{object-dimensions Cell and feature names for an Assay object}
\usage{
\method{dimnames}{Assay}(x, ...)
}
\arguments{
\item{x}{A SpatialX or Assay object}

\item{...}{Additional arguments passed to methods}
}
\description{
object-dimensions Cell and feature names for an Assay object
}
