% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialx.R
\name{dimnames.SpatialX}
\alias{dimnames.SpatialX}
\title{object-dimensions Cell and feature names for the active assay
in a SpatialX object}
\usage{
\method{dimnames}{SpatialX}(x, ...)
}
\arguments{
\item{x}{A SpatialX or Assay object}

\item{...}{Additional arguments passed to methods}
}
\description{
object-dimensions Cell and feature names for the active assay
in a SpatialX object
}
