% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastComputeQuantitiesWithoutCVT}
\alias{fastComputeQuantitiesWithoutCVT}
\title{Fast compute the testing quantities without covariates, double format
input all genes at a time
include the loop number of gene in Cpp file}
\usage{
fastComputeQuantitiesWithoutCVT(
  Yin,
  locationin,
  kernel_paramin,
  Din,
  tauin,
  kernel_typein
)
}
\arguments{
\item{Yin}{Working vector}

\item{locationin}{Location for each cell}

\item{kernel_paramin}{Gaussian kernel parameters}

\item{Din}{Weight for each gene}

\item{tauin}{Initial value for variance component}

\item{kernel_typein}{The kernel type, including 'gaussian', 'linear' and 'cosine'}
}
\value{
A list
}
\description{
Fast compute the testing quantities without covariates, double format
input all genes at a time
include the loop number of gene in Cpp file
}
