% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastTraceComputeTestQuantRcpp_cov}
\alias{FastTraceComputeTestQuantRcpp_cov}
\title{Compute the testing quantities with covariates, fast trace term}
\usage{
FastTraceComputeTestQuantRcpp_cov(Pyin, Xin, Kin, Din, tauin)
}
\arguments{
\item{Pyin}{The vector P*y}

\item{Xin}{covariates}

\item{Kin}{Kernel matrix to be tested}

\item{Din}{Weight for each gene}

\item{tauin}{Initial value for variance component}
}
\value{
A list
}
\description{
Compute the testing quantities with covariates, fast trace term
}
