% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialx.R
\name{head.SpatialX}
\alias{head.SpatialX}
\title{SpatialX-class Get the first rows of cell-level metadata}
\usage{
\method{head}{SpatialX}(x, n = 10L, ...)
}
\arguments{
\item{x}{A SpatialX object}

\item{n}{Number of rows to return}

\item{...}{Additional arguments passed to \code{\link[utils]{head}}}
}
\value{
\code{head}: The first \code{n} rows of cell-level metadata
}
\description{
SpatialX-class Get the first rows of cell-level metadata
}
\examples{
\dontrun{
# Get the first 10 rows of cell-level metadata
head(spatialx_obj)

# Get first 5 rows
head(spatialx_obj, n = 5)
}
}
