% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilties.R
\name{nb_loglik}
\alias{nb_loglik}
\alias{nb_loglik_mom}
\alias{fit_nb_optim}
\alias{pos_loglik}
\alias{fit_pos_optim}
\title{log-likelihood for two-parameter negative binomial but fix the mu}
\usage{
nb_loglik(param, y)

nb_loglik_mom(th, mu, y)

fit_nb_optim(x, maxiter = 500)

pos_loglik(mu, y)

fit_pos_optim(x, maxiter = 500)
}
\description{
log-likelihood for two-parameter negative binomial but fix the mu

fitting data with nb through optim function

log-likelihood for poisson distribution

fitting data with poisson through optim function
}
\keyword{internal}
