% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MatrixInverse}
\alias{MatrixInverse}
\title{Do inverse of sysmetric matrix}
\usage{
MatrixInverse(Min)
}
\arguments{
\item{Min}{A sysmetric matrix}
}
\value{
A list
}
\description{
Do inverse of sysmetric matrix
}
