% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialx.R
\name{[.Assay}
\alias{[.Assay}
\title{Assay-class Subset assay data}
\usage{
\method{[}{Assay}(x, i, j, ...)
}
\arguments{
\item{x}{An Assay object}

\item{i}{Feature indices or names}

\item{j}{Cell indices or names}

\item{...}{Additional arguments passed to methods}
}
\value{
\code{[}: The \code{data} slot for features \code{i} and cells \code{j}
}
\description{
Assay-class Subset assay data
}
\examples{
\dontrun{
# Subset assay by features and cells
subset_data <- assay_obj[1:10, 1:5]

# Subset by feature names
specific_genes <- assay_obj[c("Gene1", "Gene2"), ]

# Subset by cell names
specific_cells <- assay_obj[, c("Cell1", "Cell2")]
}


}
