% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialx.R
\name{tail.SpatialX}
\alias{tail.SpatialX}
\title{SpatialX-class Get the last rows of cell-level metadata}
\usage{
\method{tail}{SpatialX}(x, n = 10L, ...)
}
\arguments{
\item{x}{A SpatialX object}

\item{n}{Number of rows to return}

\item{...}{Additional arguments passed to \code{\link[utils]{tail}}}
}
\value{
\code{tail}: The last \code{n} rows of cell-level metadata
}
\description{
SpatialX-class Get the last rows of cell-level metadata
}
\examples{
\dontrun{
# Get the last 10 rows of cell-level metadata
tail(spatialx_obj)

# Get last 5 rows
tail(spatialx_obj, n = 5)
}

}
